Additional test, refactor compilation into own module
