+++ plutarch.cabal
@@ -75,6 +75,7 @@ library
     Plutarch.Backend.ANF
     Plutarch.Backend.Arity
     Plutarch.Backend.AST
+    Plutarch.Backend.Compile
     Plutarch.Backend.PosTree
     Plutarch.Backend.RawTerm
     Plutarch.Backend.Term
@@ -189,6 +190,7 @@ test-suite tests
   hs-source-dirs: test
   build-depends:
     , mtl
+    , nonempty-vector
     , plutarch
     , plutus-core
     , pretty-show
